Projection graphics.
Taking a world, objects in that world, defined by points,

and a view/screen

we can now project those points onto the screen and get their projected
value.

We haven't currently got a way of changing that basis to make the plane
(which might be wonky) into a 2d (1,0),(0,1) basis that we could put on
a screen.

We're getting there!